+++ lib/segment/src/index/field_index/map_index/mod.rs
@@ -17,7 +17,6 @@ pub mod mutable_map_index;
 pub mod on_disk_map_index;
 mod payload_index_impl;
 pub mod prefix_index;
-pub mod prefix_read;
 pub mod read_ops;
 #[cfg(test)]
 mod tests;
